requirements.txt
# Web Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.0.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
asyncpg==0.28.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# AI & ML
openai==0.28.0
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.1.0

# API Integrations
requests==2.31.0
httpx==0.24.1
aiohttp==3.8.5

# Chatbot Integration
# botpress-py==0.1.1

# Social Media Integration
# python-buffer==0.1.0
tweepy==4.14.0
facebook-sdk==3.1.0
linkedin-api==2.0.0a1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1

# Development Tools
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.0.0
pyyaml==6.0.1
Jinja2==3.1.2